moved few methods away from DataManipulation class as have nothing to do with Frames but more generic + created a separate class with Time Series specific functionality
