+++ sources/common/DataManipulator.py
@@ -1,14 +1,10 @@
 """
 Data Manipulation class
 """
-# package to handle date and times
-from datetime import timedelta
 # package to add support for multi-language (i18n)
 import gettext
 # package to handle files/folders and related metadata/operations
 import os
-# package facilitating Data Frames manipulation
-import pandas as pd
 
 
 class DataManipulator:
@@ -43,15 +39,6 @@ class DataManipulator:
         return input_data_frame
 
     @staticmethod
-    def fn_add_days_within_column_to_data_frame(input_data_frame, dict_expression):
-        input_data_frame['Days Within'] = input_data_frame[dict_expression['End Date']] - \
-                                          input_data_frame[dict_expression['Start Date']] + \
-                                          timedelta(days=1)
-        input_data_frame['Days Within'] = input_data_frame['Days Within'] \
-            .apply(lambda x: int(str(x).replace(' days 00:00:00', '')))
-        return input_data_frame
-
-    @staticmethod
     def fn_add_minimum_and_maximum_columns_to_data_frame(input_data_frame, dict_expression):
         grouped_df = input_data_frame.groupby(dict_expression['group_by']) \
             .agg({dict_expression['calculation']: ['min', 'max']})
@@ -60,89 +47,6 @@ class DataManipulator:
         if 'map' in dict_expression:
             grouped_df.rename(columns=dict_expression['map'], inplace=True)
         return grouped_df
-
-    def fn_add_timeline_evaluation_column_to_data_frame(self, in_df, dict_expression):
-        # shorten last method parameter
-        de = dict_expression
-        # add helpful column to use on "Timeline Evaluation" column determination
-        in_df['rd'] = de['Reference Date']
-        # rename some columns to cope with long expression
-        in_df.rename(columns={'Start Date': 'sd', 'End Date': 'ed'}, inplace=True)
-        # actual "Timeline Evaluation" column determination
-        cols = ['rd', 'sd', 'ed']
-        in_df['Timeline Evaluation'] = in_df[cols].apply(lambda r: 'Current'
-                                                         if r['sd'] <= r['rd'] <= r['ed'] else
-                                                         'Past' if r['sd'] < r['rd'] else 'Future',
-                                                         axis=1)
-        # rename back columns
-        in_df.rename(columns={'sd': 'Start Date', 'ed': 'End Date', 'rd': 'Reference Date'},
-                     inplace=True)
-        # decide if the helpful column is to be retained or not
-        removal_needed = self.fn_decide_by_omission_or_specific_false(de, 'Keep Reference Date')
-        if removal_needed:
-            in_df.drop(columns=['Reference Date'], inplace=True)
-        return in_df
-
-    def fn_add_value_to_dictionary(self, in_list, adding_value, adding_type, reference_column):
-        add_type = adding_type.lower()
-        total_columns = len(in_list)
-        reference_indexes = {
-            'add': {'after': 0, 'before': 0},
-            'cycle_down_to': {'after': 0, 'before': 0}
-        }
-        if type(reference_column) is int:
-            reference_indexes = {
-                'add': {
-                    'after': in_list.copy().index(reference_column) + 1,
-                    'before': in_list.copy().index(reference_column),
-                },
-                'cycle_down_to': {
-                    'after': in_list.copy().index(reference_column),
-                    'before': in_list.copy().index(reference_column),
-                }
-            }
-        positions = {
-            'after': {
-                'cycle_down_to': reference_indexes.get('cycle_down_to').get('after'),
-                'add': reference_indexes.get('add').get('after'),
-            },
-            'before': {
-                'cycle_down_to': reference_indexes.get('cycle_down_to').get('before'),
-                'add': reference_indexes.get('add').get('before'),
-            },
-            'first': {
-                'cycle_down_to': 0,
-                'add': 0,
-            },
-            'last': {
-                'cycle_down_to': total_columns,
-                'add': total_columns,
-            }
-        }
-        return self.add_value_to_dictionary_by_position({
-            'adding_value': adding_value,
-            'list': in_list,
-            'position_to_add': positions.get(add_type).get('add'),
-            'position_to_cycle_down_to': positions.get(add_type).get('cycle_down_to'),
-            'total_columns': total_columns,
-        })
-
-    @staticmethod
-    def add_value_to_dictionary_by_position(adding_dictionary):
-        list_with_values = adding_dictionary['list']
-        list_with_values.append(adding_dictionary['total_columns'])
-        for counter in range(adding_dictionary['total_columns'],
-                             adding_dictionary['position_to_cycle_down_to'], -1):
-            list_with_values[counter] = list_with_values[(counter - 1)]
-        list_with_values[adding_dictionary['position_to_add']] = adding_dictionary['adding_value']
-        return list_with_values
-
-    @staticmethod
-    def fn_add_weekday_columns_to_data_frame(input_data_frame, columns_list):
-        for current_column in columns_list:
-            input_data_frame['Weekday for ' + current_column] = input_data_frame[current_column] \
-                .apply(lambda x: x.strftime('%A'))
-        return input_data_frame
 
     def fn_apply_query_to_data_frame(self, local_logger, timmer, input_data_frame, extract_params):
         timmer.start()
@@ -175,29 +79,6 @@ class DataManipulator:
         input_data_frame.query(query_expression, inplace=True)
         timmer.stop()
         return input_data_frame
-
-    @staticmethod
-    def fn_convert_datetime_columns_to_string(input_data_frame, columns_list, columns_format):
-        for current_column in columns_list:
-            input_data_frame[current_column] = \
-                input_data_frame[current_column].map(lambda x: x.strftime(columns_format))
-        return input_data_frame
-
-    @staticmethod
-    def fn_convert_string_columns_to_datetime(input_data_frame, columns_list, columns_format):
-        for current_column in columns_list:
-            input_data_frame[current_column] = pd.to_datetime(input_data_frame[current_column],
-                                                              format=columns_format)
-        return input_data_frame
-
-    @staticmethod
-    def fn_decide_by_omission_or_specific_false(in_dictionary, key_decision_factor):
-        removal_needed = False
-        if key_decision_factor not in in_dictionary:
-            removal_needed = True
-        elif not in_dictionary[key_decision_factor]:
-            removal_needed = True
-        return removal_needed
 
     def fn_filter_data_frame_by_index(self, local_logger, in_data_frame, filter_rule):
         reference_expression = filter_rule['Query Expression for Reference Index']
@@ -233,10 +114,8 @@ class DataManipulator:
         return column_index_to_return
 
     @staticmethod
-    def fn_get_first_current_and_last_column_value_from_data_frame(in_data_frame, in_column_name):
+    def fn_get_first_and_last_column_value_from_data_frame(in_data_frame, in_column_name):
         return {
             'first': in_data_frame.iloc[0][in_column_name],
-            'current': in_data_frame.query('`Timeline Evaluation` == "Current"',
-                                           inplace=False)[in_column_name].max(),
             'last': in_data_frame.iloc[(len(in_data_frame) - 1)][in_column_name],
         }
